Hide game when loading. Updated ServerSwitcher UX
I noticed when loading the mod in a VM that it was relatively easy
to cause game to crash as you are able to interact with game as it
is being patched. Hiding the game window, while ugly, resolves that.

ServerSwitcher input box has been adjusted to be more aligned with
rest of OnlineMenu UI.
The game now also tells you what location it's using for server data